pyproject.toml
[project]
name = "lignualeo-scripts"
version = "0.1.0"
description = "Utilities and bots for importing words into LinguaLeo."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32",
    "python-telegram-bot>=21.6",
    "python-dotenv>=1.0.0",
]
